requirements.txt
# Core LLM and Embeddings
transformers>=4.25.1
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector DB (FAISS)
faiss-cpu>=1.7.4

# Document Parsing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21
markdown
csvkit

# Backend / Utility
numpy>=1.24.0
scikit-learn>=1.2.2

# Frontend
streamlit>=1.25.0

# Optional: For saving/loading models/chunks
pickle-mixin

# Avoid Streamlit errors on Windows
watchdog>=2.1.9
